+++ package.json
@@ -12,6 +12,7 @@
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.2",
     "react-scripts": "5.0.1",
+    "tailwind": "^4.0.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
