+++ src/svg.d.ts
@@ -1,4 +1,4 @@
 declare module "*.svg" {
-  const content: string;
+  const content: any;
   export default content;
 }
