+++ src/utils/enum.tsx
@@ -10,3 +10,16 @@ export const GENRES: string[] = [
     'DANCE',
     'ETC',
 ];
+
+export const GENRES_KOR: string[] = [
+    "발라드",
+    "K-POP",
+    "랩/힙합",
+    "인디",
+    "트로트",
+    "락",
+    "R&B",
+    "동요",
+    "댄스",
+    "기타"
+  ];
